graphene_workshop_completed/pyproject.toml
[tool.poetry]
name = "graphene_workshop_completed"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask-GraphQL = "^2.0.1"
graphene-sqlalchemy = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
mypy = "^0.931"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
